meta-ampere: power control: Adjust "power soft/off" the host in Mt.Jade
The current ampere-hostctl package removes the power off/on services
of phosphor-state-manager and uses ampere services. This solution is not
correct. Ampere platform should use power control functions of
phosphor-state-manager. And only appends or overides the default services
by Ampere's services if need.

By default, To handle power soft action, phosphor-state-manager will trigger
obmc-host-shutdown@0.target. This target then call
xyz.openbmc_project.Ipmi.Internal.SoftPowerOff.service to request OS shutdown.
When the host OS shutdown is already done, the target will trigger
obmc-chassis-poweroff@0.target to turn off the chassis.
The default xyz.openbmc_project.Ipmi.Internal.SoftPowerOff.service uses
Ipmi inband to communicate with the host. It is difference with Ampere Mt.Jade
platform. We uses the GPIO pin SHUTDOWN_REQUEST to request shutdown the
host OS. The host will trigger SHUTDOWN_ACK when shutdown is done.
This cause xyz.openbmc_project.Ipmi.Internal.SoftPowerOff.service will be
overide by ampere service. This service will trigger SHUTDOWN_REQUEST pin
and wait for SHUTDOWN_ACK before start obmc-chassis-poweroff@0.target.

Phosphor-state-manager will also trigger obmc-host-shutdown@0.target to
handle power off actions. But it will bypass the OS shutdown steps.

This commit removes ampere-chassis-poweroff, ampere-host-shutdown serives
which handle power off/soft actions and restore phosphor-state-manager's
services. It also supports ampere.xyz.openbmc_project.Ipmi.Internal.
SoftPowerOff.service which will overide
xyz.openbmc_project.Ipmi.Internal.SoftPowerOff.service.

Tested:
	1. Create "power soft" actions use ipmitool, redfish and BMC web.
	2. Make sure the power order will be shutdown the host then power
	of the chassis.
	3. Create "power off" actions use ipmitool, redfish and BMC web.
	4. Make sure the power action will be power off the chassis.

+++ meta-ampere/meta-common/recipes-ac01/host/files/ampere_power_util.sh
@@ -24,15 +24,11 @@ GPIO_SPI0_BACKUP_SEL=227
 GPIO_BASE=$(cat /sys/class/gpio/gpio*/base)
 
 function usage() {
-  echo "usage: power-util mb [on|off|status|cycle|reset|graceful_shutdown|graceful_reset|force_reset]";
+  echo "usage: power-util mb [on|status|cycle|reset|graceful_reset|force_reset]";
 }
 
 power_off() {
-  echo "Shutting down Server $2"
-  set_gpio_active_low $((${GPIO_BASE} + ${GPIO_OCP_AUX_PWREN})) high
-  set_gpio_active_low $((${GPIO_BASE} + ${GPIO_OCP_MAIN_PWREN})) low
-  set_gpio_active_low $((${GPIO_BASE} + ${GPIO_SPI0_PROGRAM_SEL})) low
-  busctl set-property xyz.openbmc_project.State.Chassis /xyz/openbmc_project/state/chassis0 xyz.openbmc_project.State.Chassis RequestedPowerTransition s xyz.openbmc_project.State.Chassis.Transition.Off
+  echo "power_off"
 }
 
 power_on() {
@@ -63,26 +59,31 @@ timestamp() {
 }
 
 shutdown_ack() {
-  echo "Receive shutdown ACK triggered"
-  power_off
+  if [ -f "/run/openbmc/host@0-softpoweroff" ]; then
+    echo "Receive shutdown ACK triggered after softportoff the host."
+    touch /run/openbmc/host@0-softpoweroff-shutdown-ack
+  else
+    echo "Receive shutdown ACK triggered"
+    power_off
 
-  current_time="$(timestamp)"
-  if [ -f "/run/openbmc/host@0-shutdown-req-time" ]; then
-    request_time="$(cat /run/openbmc/host@0-shutdown-req-time)"
-    # shutdow request must be done in 30s
-    if [[ "$current_time" -lt "$request_time + 30" ]]; then
-      power_off
-      sleep 5s
-      if [ -f "/run/openbmc/host@0-graceful-reset-time" ]; then
-          req_reset="$(cat /run/openbmc/host@0-graceful-reset-time)"
-          if [[ "$current_time" -lt "$req_reset + 45" ]]; then
-            echo "powering on because of reset request"
-            power_on
-            rm -rf "/run/openbmc/host@0-graceful-reset-time"
-          fi
+    current_time="$(timestamp)"
+    if [ -f "/run/openbmc/host@0-shutdown-req-time" ]; then
+      request_time="$(cat /run/openbmc/host@0-shutdown-req-time)"
+      # shutdow request must be done in 30s
+      if [[ "$current_time" -lt "$request_time + 30" ]]; then
+        power_off
+        sleep 5s
+        if [ -f "/run/openbmc/host@0-graceful-reset-time" ]; then
+            req_reset="$(cat /run/openbmc/host@0-graceful-reset-time)"
+            if [[ "$current_time" -lt "$req_reset + 45" ]]; then
+              echo "powering on because of reset request"
+              power_on
+              rm -rf "/run/openbmc/host@0-graceful-reset-time"
+            fi
+        fi
       fi
+      rm -rf "/run/openbmc/host@0-shutdown-req-time"
     fi
-    rm -rf "/run/openbmc/host@0-shutdown-req-time"
   fi
 }
 
@@ -99,6 +100,33 @@ graceful_shutdown() {
     gpioset -l 0 49=0
     sleep 30s
   fi
+}
+
+soft_off() {
+  # Trigger shutdown_req
+  touch /run/openbmc/host@0-softpoweroff
+  gpioset -l 0 49=1
+  sleep 1s
+  gpioset -l 0 49=0
+
+  # Wait for shutdown_ack from the host in 30 seconds
+  cnt=30
+  while [ $cnt -gt 0 ];
+  do
+    # Wait for SHUTDOWN_ACK and create the host@0-softpoweroff-shutdown-ack
+    if [ -f "/run/openbmc/host@0-softpoweroff-shutdown-ack" ]; then
+      break
+    fi
+    sleep 1
+    cnt=$((cnt - 1))
+  done
+  # Softpoweroff is successed
+  sleep 2
+  rm -rf /run/openbmc/host@0-softpoweroff
+  if [ -f "/run/openbmc/host@0-softpoweroff-shutdown-ack" ]; then
+    rm -rf /run/openbmc/host@0-softpoweroff-shutdown-ack
+  fi
+  echo 0
 }
 
 graceful_reset() {
@@ -147,12 +175,6 @@ if [ $2 = "on" ]; then
     rm -rf /run/openbmc/host@*-on
     power_on
   fi
-elif [ $2 = "off" ]; then
-  if [ $(power_status) == "on" ]; then
-    power_off
-  else
-    rm -rf /run/openbmc/host@0-request
-  fi
 elif [ $2 == "cycle" ]; then
   if [ $(power_status) == "on" ]; then
     echo "Powering off server"
@@ -168,10 +190,6 @@ elif [ $2 == "reset" ]; then
   else
     echo "ERROR: Server not powered on"
   fi
-elif [[ $2 == "graceful_shutdown" ]]; then
-  if [ $(power_status) == "on" ]; then
-    graceful_shutdown
-  fi
 elif [ $2 == "graceful_reset" ]; then
   if [ $(power_status) == "on" ]; then
       graceful_reset
@@ -182,6 +200,14 @@ elif [ $2 == "status" ]; then
   power_status
 elif [ $2 == "force_reset" ]; then
   force_reset
+elif [ $2 == "soft_off" ]; then
+  ret=$(soft_off)
+  if [ $ret == 0 ]; then
+    echo "The host is already softoff"
+  else
+    echo "Failed to softoff the host"
+  fi
+  exit $ret;
 else
   echo "Invalid parameter2=$2"
   usage;
